refactor: .json files called from world_assets; copies of global.json removed
